requirements.txt
# Core Dependencies
pandas
gspread
oauth2client
gdown
python-pptx
pytesseract
Pillow
ollama
openpyxl
streamlit
python-dotenv
numpy
requests
pymupdf
pdfplumber
sentence-transformers
